Format files

+++ test/test.h
@@ -23,20 +23,20 @@ typedef struct Test {
         }                    \
     } while (0)
 
-#define RUN_ALL_TESTS(tests, name)                                       \
-    do {                                                                 \
-        fprintf(stderr, "[==========] Running " name " tests\n");        \
-        for (int i = 0; i < sizeof(tests) / sizeof(Test*); i++) {        \
-            bool result = true;                                          \
-            tests[i]->func(&result);                                     \
-            if (result) {                                                \
-                fprintf(stderr, "[\x1b[38;5;2m  PASSED  \x1b[0m] %s\n",  \
-                        tests[i]->desc);                                 \
-            } else {                                                     \
-                fprintf(stderr, "[\x1b[38;5;1m  FAILED  \x1b[0m] %s\n",  \
-                        tests[i]->desc);                                 \
-            }                                                            \
-        }                                                                \
+#define RUN_ALL_TESTS(tests, name)                                      \
+    do {                                                                \
+        fprintf(stderr, "[==========] Running " name " tests\n");       \
+        for (size_t i = 0; i < sizeof(tests) / sizeof(Test*); i++) {    \
+            bool result = true;                                         \
+            tests[i]->func(&result);                                    \
+            if (result) {                                               \
+                fprintf(stderr, "[\x1b[38;5;2m  PASSED  \x1b[0m] %s\n", \
+                        tests[i]->desc);                                \
+            } else {                                                    \
+                fprintf(stderr, "[\x1b[38;5;1m  FAILED  \x1b[0m] %s\n", \
+                        tests[i]->desc);                                \
+            }                                                           \
+        }                                                               \
     } while (0)
 
 #endif
